requirements.txt
flask>=2.3.0
python-dotenv
groq
